extract vmap and if
